notification renaming + settings cleanup
