+++ package.json
@@ -41,6 +41,7 @@
 		"react-native-reanimated": "~1.7.0",
 		"react-native-safe-area-context": "0.7.3",
 		"react-native-screens": "~2.2.0",
+		"react-native-shapes": "^0.1.0",
 		"react-native-vector-icons": "^6.6.0",
 		"react-native-web": "~0.11.7",
 		"react-navigation": "^4.3.9",
